Add Gitcord workflow, remove old workflows
Note: I am keeping the archival workflow in main until all of the legacy issue channels have been properly archived.
