--- .github/workflows/create_issue_channel.yaml
@@ -1,21 +1,0 @@
-name: Create Issue Channel on Discord
-
-on:
-  issues:
-    types: [opened, reopened]
-
-jobs:
-  create_channel:
-    runs-on: ubuntu-latest
-    if: ${{ !contains(github.event.issue.labels.*.name, 'automation:officer') }}
-    steps:
-      - uses: actions/checkout@v2
-      - uses: actions/setup-node@v2
-        with:
-          node-version: '16'
-      - run: npm ci
-      - run: node scripts/create-issue-channel.js ${{ github.event.issue.number }}
-        env:
-          DISCORD_BOT_TOKEN: ${{ secrets.DISCORD_BOT_TOKEN }}
-          GUILD_ID: ${{ secrets.GUILD_ID }}
-          HUB_ID: ${{ secrets.HUB_ID }}
