--- .github/workflows/create_message.yaml
@@ -1,23 +1,0 @@
-name: Create PR comment
-
-on:
-  issue_comment:
-    types:
-      - created
-
-jobs:
-  issue_commented:
-    name: Issue comment
-    if: ${{ !(github.event.issue.pull_request) && (github.event.issue.state == 'open') }}
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-      - uses: actions/setup-node@v2
-        with:
-          node-version: '16'
-      - run: npm ci
-      - run: node scripts/create-message.js ${{ github.event.issue.number }}
-        env:
-          DISCORD_BOT_TOKEN: ${{ secrets.DISCORD_BOT_TOKEN }}
-          MESSAGE: ${{ github.event.comment.body }}
-          GUILD_ID: ${{ secrets.GUILD_ID }}
