+++ .github/workflows/gitcord.yaml
@@ -1,0 +1,73 @@
+# Gitcord workflow copied from:
+# https://etok.codes/gitcord/blob/main/.github/workflows/gitcord.yaml
+name: Gitcord
+
+on:
+  issues:
+    types:
+      - opened
+      - edited
+      - deleted
+      - transferred
+      - closed
+      - reopened
+      - assigned
+      - unassigned
+      - labeled
+      - unlabeled
+      - locked
+      - unlocked
+      - milestoned
+      - demilestoned
+  issue_comment:
+    types:
+      - created
+      - edited
+      - deleted
+  pull_request:
+    types:
+      - opened
+      - edited
+      - deleted
+      - transferred
+      - review_requested
+      - review_request_removed
+      - ready_for_review
+      - closed
+      - reopened
+      - assigned
+      - unassigned
+      - labeled
+      - unlabeled
+      - locked
+      - unlocked
+      - milestoned
+      - demilestoned
+  pull_request_review:
+    types:
+      - submitted
+      - edited
+      - dismissed
+  pull_request_review_comment:
+    types:
+      - created
+      - edited
+      - deleted
+
+jobs:
+  gitcord:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Read GitHub event payload
+        id: event_payload
+        run: |
+          echo ::set-output name=payload::$(cat $GITHUB_EVENT_PATH)
+
+      - name: 'Gitcord ${{ github.event_name }}:${{ github.event.action }} (${{ github.repository }})'
+        uses: ethanthatonekid/gitcord@main
+        with:
+          github-event-name: ${{ github.event_name }}
+          github-event-payload: ${{ steps.event_payload.outputs.payload }}
+          github-token: ${{ github.token }}
+          discord-token: ${{ secrets.DISCORD_TOKEN }}
+          discord-channel-id: ${{ secrets.DISCORD_CHANNEL_ID }}
